perf(dashboard): lazy-load usage analytics charts (#4466)
Thanks @KooshaPari! Rebased onto release/v3.8.32. Recharts now lazy-loads (public analytics API preserved, source-guard test included).
