+++ src/shared/components/analytics/chartColors.ts
@@ -1,0 +1,12 @@
+export const PROVIDER_COLORS = [
+  "#f59e0b",
+  "#ef4444",
+  "#8b5cf6",
+  "#10b981",
+  "#06b6d4",
+  "#ec4899",
+  "#f97316",
+  "#6366f1",
+  "#14b8a6",
+  "#a855f7",
+];
