+++ src/shared/components/analytics/index.tsx
@@ -13,20 +13,16 @@ export {
   StatCard,
   CompactStatGrid,
   ActivityHeatmap,
-  DailyTrendChart,
-  AccountDonut,
-  ApiKeyDonut,
   ApiKeyTable,
-  WeeklyPattern,
   MostActiveDay7d,
   WeeklySquares7d,
   ModelTable,
   UsageDetail,
-  ProviderCostDonut,
-  ModelOverTimeChart,
   ProviderTable,
   ServiceTierBreakdown,
 } from "./charts";
+export { AccountDonut, ApiKeyDonut, ProviderCostDonut } from "./rechartsDonuts";
+export { DailyTrendChart, ModelOverTimeChart } from "./rechartsUsageCharts";
 
 export { default as ApiKeyFilterDropdown } from "./ApiKeyFilterDropdown";
 export { default as CustomRangePicker } from "./CustomRangePicker";
